added scaffold from spec
