--- mock_spec.yaml
@@ -1,28 +1,0 @@
-version: "1.0.0"
-id: "example_service"
-name: "Example Service"
-author:
-  name: "Heber Silva"
-tags:
-  - "tag1"
-  - "tag2"
-status: "stable"
-signals:
-  - name: "Signal 1"
-    type: "Type 1"
-    path: "/path/to/signal/1"
-  - name: "Signal 2"
-    type: "Type 2"
-    path: "/path/to/signal/2"
-images:
-  - name: "Image 1"
-    description: "Description for Image 1"
-    path: "/path/to/image/1"
-changelog:
-  path: "/path/to/changelog"
-  changes:
-    - description: "Initial release"
-      date: "2023-06-01"
-components:
-  - name: "cassandra"
-  - name: "aws_rds"
